Migrate integrations sms and spotify to new structure
