--- app/integrations/server/index.js
@@ -1,16 +1,0 @@
-import '../lib/rocketchat';
-import './logger';
-import './lib/validation';
-import './methods/incoming/addIncomingIntegration';
-import './methods/incoming/updateIncomingIntegration';
-import './methods/incoming/deleteIncomingIntegration';
-import './methods/outgoing/addOutgoingIntegration';
-import './methods/outgoing/updateOutgoingIntegration';
-import './methods/outgoing/replayOutgoingIntegration';
-import './methods/outgoing/deleteOutgoingIntegration';
-import './methods/clearIntegrationHistory';
-import './api/api';
-import './lib/triggerHandler';
-import './triggers';
-
-export { mountIntegrationQueryBasedOnPermissions, mountIntegrationHistoryQueryBasedOnPermissions } from './lib/mountQueriesBasedOnPermission';
