--- app/integrations/server/logger.js
@@ -1,8 +1,0 @@
-import { Logger } from '../../logger';
-
-export const logger = new Logger('Integrations', {
-	sections: {
-		incoming: 'Incoming WebHook',
-		outgoing: 'Outgoing WebHook',
-	},
-});
