+++ app/livechat/imports/server/rest/sms.js
@@ -5,7 +5,7 @@ import { Random } from 'meteor/random';
 import { FileUpload } from '../../../../file-upload/server';
 import { LivechatRooms, LivechatVisitors, LivechatDepartment } from '../../../../../server/models';
 import { API } from '../../../../../server/api/v1';
-import { SMS } from '../../../../sms';
+import { SMS } from '../../../../../server/integrations/sms';
 import { Livechat } from '../../../server/lib/Livechat';
 
 const getUploadFile = (details, fileUrl) => {
