--- app/sms/server/index.js
@@ -1,6 +1,0 @@
-import './settings';
-import './services/twilio';
-import './services/voxtelesys';
-import './services/mobex';
-
-export { SMS } from './SMS';
