+++ server/api/methods/index.js
@@ -8,3 +8,11 @@ import './messages/unreadMessages';
 import './messages/snippetMessage';
 import './messages/starMessage';
 import './messages/pinMessage';
+import './integrations/clearIntegrationHistory';
+import './integrations/incoming/addIncomingIntegration';
+import './integrations/incoming/deleteIncomingIntegration';
+import './integrations/incoming/updateIncomingIntegration';
+import './integrations/outgoing/addOutgoingIntegration';
+import './integrations/outgoing/deleteOutgoingIntegration';
+import './integrations/outgoing/replayOutgoingIntegration';
+import './integrations/outgoing/updateOutgoingIntegration';
