+++ server/api/methods/integrations/clearIntegrationHistory.js
@@ -1,0 +1,29 @@
+import { Meteor } from 'meteor/meteor';
+
+import { hasPermission } from '../../../../app/authorization';
+import { IntegrationHistory, Integrations } from '../../../models';
+import notifications from '../../../../app/notifications/server/lib/Notifications';
+
+Meteor.methods({
+	clearIntegrationHistory(integrationId) {
+		let integration;
+
+		if (hasPermission(this.userId, 'manage-outgoing-integrations') || hasPermission(this.userId, 'manage-outgoing-integrations', 'bot')) {
+			integration = Integrations.findOne(integrationId);
+		} else if (hasPermission(this.userId, 'manage-own-outgoing-integrations') || hasPermission(this.userId, 'manage-own-outgoing-integrations', 'bot')) {
+			integration = Integrations.findOne(integrationId, { fields: { '_createdBy._id': this.userId } });
+		} else {
+			throw new Meteor.Error('not_authorized', 'Unauthorized', { method: 'clearIntegrationHistory' });
+		}
+
+		if (!integration) {
+			throw new Meteor.Error('error-invalid-integration', 'Invalid integration', { method: 'clearIntegrationHistory' });
+		}
+
+		IntegrationHistory.removeByIntegrationId(integrationId);
+
+		notifications.streamIntegrationHistory.emit(integrationId, { type: 'removed' });
+
+		return true;
+	},
+});
