+++ server/importPackages.js
@@ -1,5 +1,4 @@
 import '../app/cors/server';
-import '../app/sms';
 import '../app/analytics/server';
 import '../app/assets/server';
 import '../app/authorization';
@@ -20,7 +19,6 @@ import '../app/federation/server';
 import '../app/file';
 import '../app/file-upload';
 import '../app/iframe-login/server';
-import '../app/integrations/server';
 import '../app/issuelinks/server';
 import '../app/katex/server';
 import '../app/ldap/server';
@@ -38,7 +36,6 @@ import '../app/oembed/server';
 import '../app/otr/server';
 import '../app/push-notifications/server';
 import '../app/retention-policy';
-import '../app/spotify/server';
 import '../app/theme/server';
 import '../app/threads/server';
 import '../app/tokenpass/server';
