+++ server/integrations/custom/index.js
@@ -1,0 +1,8 @@
+import './rocketchat';
+import './logger';
+import './lib/validation';
+import './api/api';
+import './lib/triggerHandler';
+import './triggers';
+
+export { mountIntegrationQueryBasedOnPermissions, mountIntegrationHistoryQueryBasedOnPermissions } from './lib/mountQueriesBasedOnPermission';
