+++ server/integrations/custom/logger.js
@@ -1,0 +1,8 @@
+import { Logger } from '../../../app/logger';
+
+export const logger = new Logger('Integrations', {
+	sections: {
+		incoming: 'Incoming WebHook',
+		outgoing: 'Outgoing WebHook',
+	},
+});
