+++ server/integrations/index.ts
@@ -2,3 +2,6 @@ import './slackbridge';
 import './smarsh-connector';
 import './irc';
 import './google-vision';
+import './spotify';
+import './sms';
+import './custom';
