+++ server/integrations/sms/index.js
@@ -1,0 +1,5 @@
+import './services/twilio';
+import './services/voxtelesys';
+import './services/mobex';
+
+export { SMS } from './SMS';
