+++ server/integrations/spotify/index.js
@@ -1,0 +1,10 @@
+import { Meteor } from 'meteor/meteor';
+
+import { callbacks } from '../../../app/callbacks';
+import { createSpotifyBeforeSaveMessageHandler } from './spotify';
+
+Meteor.startup(() => {
+	const beforeSaveMessage = createSpotifyBeforeSaveMessageHandler();
+
+	callbacks.add('beforeSaveMessage', beforeSaveMessage, callbacks.priority.LOW, 'spotify-save');
+});
