+++ server/services/meteor/service.ts
@@ -16,7 +16,7 @@ import { RoutingManager } from '../../../app/livechat/server/lib/RoutingManager'
 import { onlineAgents, monitorAgents } from '../../../app/livechat/server/lib/stream/agentStatus';
 import { IUser } from '../../../definition/IUser';
 import { matrixBroadCastActions } from '../../stream/streamBroadcast';
-import { integrations } from '../../../app/integrations/server/lib/triggerHandler';
+import { integrations } from '../../integrations/custom/lib/triggerHandler';
 import { ListenersModule, minimongoChangeMap } from '../../modules/listeners/listeners.module';
 import notifications from '../../../app/notifications/server/lib/Notifications';
 import { configureEmailInboxes } from '../../features/EmailInbox/EmailInbox';
